upcoming-features/kyc-backend-app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
email-validator==2.1.0
boto3==1.34.0
python-dotenv==1.0.0
typing-extensions>=4.12.0
anyio==3.7.1
sniffio==1.3.0
idna==3.6 